+++ paddleformers/transformers/auto/configuration.py
@@ -34,7 +34,6 @@ __all__ = [
 CONFIG_MAPPING_NAMES = OrderedDict(
     [
         ("bert", "BertConfig"),
-        ("deepseek_v2", "DeepseekV2Config"),
         ("deepseek_v3", "DeepseekV3Config"),
         ("ernie4_5", "Ernie4_5Config"),
         ("ernie4_5_moe", "Ernie4_5_MoeConfig"),
