+++ paddleformers/transformers/auto/modeling.py
@@ -53,7 +53,6 @@ __all__ = [
 MAPPING_NAMES = OrderedDict(
     [
         ("Bert", "bert"),
-        ("DeepseekV2", "deepseek_v2"),
         ("DeepseekV3", "deepseek_v3"),
         ("Ernie4_5", "ernie4_5"),
         ("Ernie4_5_Moe", "ernie4_5_moe"),
